COLORADO-1197 - fixed searching in RELS-INT based on non-existent PID
I am not sure if this code is ever called but as it is a copy of the
incorrect code in adr_basic_viewer I have changed it here as well.

The reference ot the PID of the object being searched for a thumbnail
was referenced by variable that was incorrect.  The variable used was
this->pid (did not exist) and should have been this->item.pid.